requirements.txt
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
python-decouple==3.8
psycopg2-binary==2.9.7
